% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{auto.arima_fore}
\alias{auto.arima_fore}
\title{The forecasting model ARIMA basd on \code{forecast::auto.arima}}
\usage{
auto.arima_fore(x, train_h, PI_level)
}
\arguments{
\item{x}{The input time series.}

\item{train_h}{The amount of future time steps to forecast.}

\item{PI_level}{Predictive Interval level, used to extract out-of-sample variance from forecasting models.}
}
\value{
A list including forecasting mean and sd.
}
\description{
The forecasting model ARIMA basd on \code{forecast::auto.arima}
}
