% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{egarch_fore}
\alias{egarch_fore}
\title{EGARCH forecasting model based on the \code{rugarch} package}
\usage{
egarch_fore(x, train_h, PI_level)
}
\arguments{
\item{x}{The input time series.}

\item{train_h}{The amount of future time steps to forecast.}

\item{PI_level}{This function does not use it in the calculation.
This is used to make this function consistent with the parameter setting of the top four models.}
}
\value{
A list including forecasting mean and sd.
}
\description{
EGARCH forecasting model based on the \code{rugarch} package
}
