% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ets_fore}
\alias{ets_fore}
\title{The forecasting model ETS basd on \code{forecast::ets}}
\usage{
ets_fore(x, train_h, PI_level)
}
\arguments{
\item{x}{The input time series.}

\item{train_h}{The amount of future time steps to forecast.}

\item{PI_level}{Predictive Interval level, used to extract out-of-sample variance from forecasting models.}
}
\value{
A list including forecasting mean and sd.
}
\description{
The forecasting model ETS basd on \code{forecast::ets}
}
