% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{febama_mcmc}
\alias{febama_mcmc}
\title{Inference procedure of FEBAMA framework}
\usage{
febama_mcmc(data, model_conf)
}
\arguments{
\item{data}{A list with \code{lpd} and \code{feat}
(the output of function \code{lpd_features_multi}).}

\item{model_conf}{Parameter settings of FEBAMA framework. Defualt \code{model_conf_default()}.}
}
\value{
\code{febama_mcmc} returns a list with the entries:
\describe{
\item{beta}{A list of (number of models -1) matrices of beta in every iteration. }
\item{betaIdx}{A list of (number of models -1) matrices of betaIdx in every iteration.}
\item{accept_prob}{A list of (number of models -1) matrices of accept probabilities in every iteration.}
}
}
\description{
In the inference procedure, we take the MAP estimation with the standard BFGS algorithm.
If variable selection is considered (\code{model_conf$algArgs$nIter > 1}) simultaneously with MAP,
we utilize the Gibbs sampler to perform variable selection over all forecasting models.
}
