% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_feature_results_multi}
\alias{forecast_feature_results_multi}
\title{Forecasting with FEBAMA}
\usage{
forecast_feature_results_multi(
  ts,
  model_conf = model_conf_curr,
  intercept = T,
  data,
  beta_out
)
}
\arguments{
\item{ts}{A list with related information of a time series:
\describe{
\item{x}{Historical data.}
\item{xx}{Real data in the forecasting horizon.}
}}

\item{model_conf}{Parameter settings of FEBAMA framework. Defualt \code{model_conf_default()}.}

\item{data}{A list with
\describe{
\item{lpd}{Log probability densities from historical data.}
\item{feat}{Features from historical data.}
\item{feat_mean}{The mean of features.}
\item{feat_sd}{The sd of features.}
}}

\item{beta_out}{The output of function \code{febama_mcmc}.}
}
\value{
\code{forecast_feature_results_multi} returns a list with the entries:
\describe{
\item{err_feature}{Log score, mase, smape measures of forecasts.}
\item{ff_feature}{Forecast values in the forecasting horizon.}
\item{w_time_varying}{Time-varying weights in the forecasting horizon.}
}
}
\description{
Forecast based on FEBAMA framework and obtain error measures.
}
