% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{log_posterior}
\alias{log_posterior}
\title{Calculate the log posterior}
\usage{
log_posterior(
  data,
  beta,
  betaIdx,
  priArgs,
  varSelArgs,
  features_used,
  model_update = 1:length(betaIdx)
)
}
\arguments{
\item{data}{A list with \code{lpd} and \code{feat}
(the output of function \code{lpd_features_multi}).}

\item{beta}{A list of coefficient vectors of the features.}

\item{betaIdx}{A list of indicator vectors.}

\item{priArgs}{Parameter settings in the priors of \code{beta} and \code{betaIdx}.}

\item{varSelArgs}{Variable selection settings.}

\item{features_used}{The features used for forecast combination.}
}
\value{
\code{log_posterior} returns the value of log posterior.
}
\description{
Calculate the log posterior of  unknown parameters in FEBAMA framework.
See \code{model_conf_default} for more information of parameter settings.
}
