% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{log_priors}
\alias{log_priors}
\title{Calculate the log joint prior}
\usage{
log_priors(beta, betaIdx, varSelArgs, priArgs, sum = TRUE)
}
\arguments{
\item{beta}{A list of coefficient vectors of the features.}

\item{betaIdx}{A list of indicator vectors.}

\item{varSelArgs}{Variable selection settings.}

\item{priArgs}{Parameter settings in the priors of \code{beta} and \code{betaIdx}.}

\item{sum}{Logical, if TRUE, sum all conditional priors.}
}
\value{
\code{log_priors} returns the value of log joint prior.
}
\description{
Calculate the log joint prior of  indicator vectors and  coefficient vectors.
See \code{model_conf_default} for more information of parameter settings.
}
