% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logscore.R
\name{logscore}
\alias{logscore}
\title{Calculate the log predictive score}
\usage{
logscore(data, beta, betaIdx, features_used, sum = TRUE)
}
\arguments{
\item{data}{A list with \code{lpd} and \code{feat}
(the output of function \code{lpd_features_multi}).}

\item{beta}{A list of coefficient vectors of the features.}

\item{betaIdx}{A list of indicator vectors.}

\item{features_used}{The features used for forecast combination.
See the parameter seting for more information.}

\item{sum}{If TRUE, return the sume of log predictive densities.}
}
\value{
\code{logscore} returns the value of log predictive score.
}
\description{
Calculate the log predictive score of a time series for FEBAMA framework.
The weights in forecast combination are related to time series features.
}
