% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities_features.R
\name{lpd_features_multi}
\alias{lpd_features_multi}
\title{Compute log predictive densities and features}
\usage{
lpd_features_multi(data, model_conf)
}
\arguments{
\item{data}{A list with related information for a time series. Historical data \code{x} is required.}

\item{model_conf}{Parameter settings for the FEBAMA framework. Defaults can be created by \code{model_conf_default()}.}
}
\value{
A list with log predictive densities and features.
}
\description{
Compute log predictive densities and features for training in FEBAMA framework.
}
