% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{model_conf_default}
\alias{model_conf_default}
\title{Default parameter settings for the FEBAMA framework}
\usage{
model_conf_default()
}
\value{
A list of FEBAMA model, prior, and algorithm settings.
}
\description{
Default parameter settings for the FEBAMA framework
}
