% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{summarize_performance}
\alias{summarize_performance}
\title{Summarize FEBAMA forecast performance}
\usage{
summarize_performance(forecasts)
}
\arguments{
\item{forecasts}{A forecast result from \code{\link[=forecast_febama]{forecast_febama()}} or a list of them.}
}
\value{
Matrix with log score, MASE, and SMAPE.
}
\description{
Summarize FEBAMA forecast performance
}
