% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities_features.R
\name{feature_clean}
\alias{feature_clean}
\title{Delete the features with NaN and add attributes}
\usage{
feature_clean(lpd_features)
}
\arguments{
\item{lpd_features}{A list of several outputs of \code{lpd_features_multi}.}
}
\value{
A list with log predictive densities and features.
}
\description{
Delete the features with NaN and add attributes
}
