% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{model_conf_default}
\alias{model_conf_default}
\title{Defualt parameter settings of FEBAMA framework}
\usage{
model_conf_default()
}
\description{
Defualt parameter settings of FEBAMA framework
}
